Just commit what we have
I didn't manage to get DuoAI to work (i.e. produce invariants) with
the traces we were generating from mypyvy. The strong suspicion is that
this is an issue in DuoAI itself, since our trace files were subsets of
those produced by the simulation -- so in the worst case, it should have
taken DuoAI longer to produce/discover the invariant.

Instead, it simply finished whilst failing to find an invariant -- which
shouldn't happen if the arguments in the paper are true. (They are not,
at least for the implementation, and it seems to some extent, not true
for the bottom-up algo itself.)